Created custom high level Callback abstract class that implements retrofit.Callback.
With this I can log actual network errors to Crashlytics with Timber.e() automatically. When there is no internet connection in a device, we are logging these. That's why we have Level 5 network errors in Crashlytics. Those should be logged only with debug log.

Moreover, when there is an internet connection error. It is passed to the caller so that they can show Snacker with appropriate network error.