FirstStartStep2 refactored.
GoogleApiClient responsibility of FirstStep2Fragment is removed. It is now just a small simple Fragment that sends Click events to the Activity.
Activity handles recreation of GoogleApiClient and every thing.

setCurrentItem(1, true) is now setCurrentItem(1) because the default is true anyways.

Opening MainActivity is moved from finish() to proper place.