Unused SearchActivity and GdgProvider removed.
This SearchAcitivity and its provider were not used inside the app. This SearchActivity was crashing, nobody was able to open it.

It could be opened by really really old Android devices with no GMS. In the past, you could include your app inside global search by doing this. New Google search removed this functionality in favour of AppIndexing.